HTML Element Update
Adds the base functionality needed to allow interactions between game entities and html elements. Note: Due to the current unfinished state of the arena pages this was built out to work in the temporary Test-arena.html file. Move work will need to be done to integrate this functionality with the completed arena pages.