Added generic pickup function
The player can now press f to pickup any destructible entities that they are on top of. They can only pick up one entity and will drop the entity if f is pressed again. There is not currently an animation for the entity being picked up but the core functionality works.